+++ filament/src/details/Material.h
@@ -69,6 +69,8 @@ public:
 
     bool hasParameter(const char* name) const noexcept;
 
+    UniformInterfaceBlock::UniformInfo const* reflect(utils::StaticString const& name) const noexcept;
+
     FMaterialInstance const* getDefaultInstance() const noexcept { return &mDefaultInstance; }
     FMaterialInstance* getDefaultInstance() noexcept { return &mDefaultInstance; }