+++ filament/src/fg/FrameGraph.h
@@ -180,18 +180,18 @@ public:
         return *pass;
     }
 
-    // Adds a reference to 'input', preventing it from being culled.
-    void present(FrameGraphHandle input);
-
     // Adds a simple execute-only pass with side effect (so it's not culled)
     template<typename Execute>
-    void simpleSideEffectPass(const char* name, Execute&& execute) {
-        addPass<Empty>(name, [](FrameGraph::Builder& builder, auto& data) { builder.sideEffect(); },
-                [execute](FrameGraphPassResources const& resources, auto const& data,
+    void addTrivialSideEffectPass(const char* name, Execute&& execute) {
+        addPass<Empty>(name, [](FrameGraph::Builder& builder, auto&) { builder.sideEffect(); },
+                [execute](FrameGraphPassResources const&, auto const&,
                         backend::DriverApi& driver) {
                     execute();
                 });
     }
+
+    // Adds a reference to 'input', preventing it from being culled.
+    void present(FrameGraphHandle input);
 
     // Returns whether the resource handle is valid. A resource handle becomes invalid after
     // it's used to declare a resource write (see Builder::write()).