+++ libs/filabridge/src/SibGenerator.cpp
@@ -36,6 +36,7 @@ SamplerInterfaceBlock const& SibGenerator::getPerViewSib() noexcept {
             .add("iblDFG",        Type::SAMPLER_2D,      Format::FLOAT, Precision::MEDIUM)
             .add("iblSpecular",   Type::SAMPLER_CUBEMAP, Format::FLOAT, Precision::MEDIUM)
             .add("ssao",          Type::SAMPLER_2D,      Format::FLOAT, Precision::MEDIUM)
+            .add("ssr",           Type::SAMPLER_2D,      Format::FLOAT, Precision::MEDIUM)
             .build();
 
     assert(sib.getSize() == PerViewSib::SAMPLER_COUNT);