+++ shaders/src/light_indirect.fs
@@ -441,8 +441,16 @@ void applyRefraction(const PixelParams pixel,
     vec3 Ft = prefilteredRadiance(ray.direction, perceptualRoughness) * frameUniforms.iblLuminance;
 #else
     // compute the point where the ray exits the medium, if needed
-    vec4 p = vec4(frameUniforms.clipFromWorldMatrix * ray.position, 1.0);
-    vec3 Ft = vec3(0.0);    // TODO: sample screen-space at p
+    vec4 p = vec4(frameUniforms.clipFromWorldMatrix * vec4(ray.position, 1.0));
+    p.xy = uvToRenderTargetUV(p.xy * (0.5 / p.w) + 0.5);
+
+    // perceptualRoughness to LOD
+    const float kNumRoughnessLods = 5.0;
+    // Empirical factor to compensate for the gaussian approximation of Dggx, chosen so
+    // cubemap and screen-space modes match at perceptualRoughness 0.125
+    float tweakedPerceptualRoughness = perceptualRoughness * 1.74;
+    float lod = log2(tweakedPerceptualRoughness * 2.0) + (kNumRoughnessLods - 1.0);
+    vec3 Ft = textureLod(light_ssr, p.xy, lod).rgb;
 #endif
 
     /* fresnel from the first interface */