+++ RELEASE_NOTES.md
@@ -6,6 +6,7 @@ A new header is inserted each time a *tag* is created.
 ## Next release
 
 - The Android support libraries (gltfio and filament-utils) now use dynamic linking.
+- Screen-space refraction is now supported.
 
 ## v1.4.5